refactor: extract IssueExportSupport.swift from IssueExportController.swift
Extracts 3 support types (IssueExportPreflightFailure error, IssueExportRequestContext,
IssueExportCompletion) from IssueExportController.swift into a new IssueExportSupport.swift.

Byte-identical move; SHA-256 7d4f5948c2fca30e8f127b233da200c07c940e250375422550235a14bac0b2e0
verified against origin/main:IssueExportController.swift lines 539-565.

Reduces IssueExportController.swift from 565 lines to 538 lines.

Tests: 667 (unchanged).
